modify: create useTheme custom hook
